+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/enumeration/IndexedEnum.java
@@ -1,0 +1,17 @@
+package com.dwarfeng.subgrade.sdk.enumeration;
+
+/**
+ * 拥有序号的枚举。
+ *
+ * @author DwArFeng
+ * @since 1.1.5
+ */
+public interface IndexedEnum {
+
+    /**
+     * 获取当前枚举的序号。
+     *
+     * @return 当前枚举的序号。
+     */
+    int getIndex();
+}