+++ subgrade-sdk/src/main/java/com/dwarfeng/subgrade/sdk/enumeration/LabeledEnum.java
@@ -1,0 +1,17 @@
+package com.dwarfeng.subgrade.sdk.enumeration;
+
+/**
+ * 拥有标签的枚举。
+ *
+ * @author DwArFeng
+ * @since 1.1.5
+ */
+public interface LabeledEnum {
+
+    /**
+     * 获取当前枚举的标签。
+     *
+     * @return 当前枚举的标签。
+     */
+    String getLabel();
+}